feat: expose `Collector` for cleaning up transient metrics
If your metrics last the whole lifetime of your service, you can just bury this.
However there are times in which they do not last that long---for example, if
there is some a startup sequence your service goes through that you need
visibility into---in which case you need a `Collector` exposed to unregister.